requirements.txt
pypdf
pdf2image
pillow
requests
openpyxl
pandas
google-genai